+++ pybricks/common/pb_type_lightmatrix.c
@@ -70,21 +70,21 @@ STATIC mp_obj_t common_LightMatrix_char(size_t n_args, const mp_obj_t *pos_args,
         pb_assert(PBIO_ERROR_INVALID_ARG);
     }
 
-    // Pick corresponding image and display it
+    // Pick corresponding icon and display it
     pb_assert(pbio_light_matrix_set_rows(self->light_matrix, pb_font_5x5[text[0] - 32]));
 
     return mp_const_none;
 }
 STATIC MP_DEFINE_CONST_FUN_OBJ_KW(common_LightMatrix_char_obj, 1, common_LightMatrix_char);
 
-static void common_LightMatrix_image__extract(mp_obj_t image_in, size_t size, uint8_t *data) {
+static void common_LightMatrix_icon__extract(mp_obj_t icon_in, size_t size, uint8_t *data) {
 
     #if MICROPY_PY_BUILTINS_FLOAT
-    // If image is a matrix, copy data from there
-    if (mp_obj_is_type(image_in, &pb_type_Matrix)) {
+    // If icon is a matrix, copy data from there
+    if (mp_obj_is_type(icon_in, &pb_type_Matrix)) {
         for (size_t r = 0; r < size; r++) {
             for (size_t c = 0; c < size; c++) {
-                float scalar = pb_type_Matrix_get_scalar(image_in, r, c);
+                float scalar = pb_type_Matrix_get_scalar(icon_in, r, c);
                 scalar = scalar > 100 ? 100 : (scalar < 0 ? 0: scalar);
                 data[r * size + c] = (uint8_t)scalar;
             }
@@ -96,7 +96,7 @@ static void common_LightMatrix_image__extract(mp_obj_t image_in, size_t size, ui
     // Unpack the main list of rows and get the requested sizes
     mp_obj_t *row_objs, *scalar_objs;
     size_t m;
-    mp_obj_get_array(image_in, &m, &row_objs);
+    mp_obj_get_array(icon_in, &m, &row_objs);
     if (m != size) {
         pb_assert(PBIO_ERROR_INVALID_ARG);
     }
@@ -115,23 +115,23 @@ static void common_LightMatrix_image__extract(mp_obj_t image_in, size_t size, ui
     }
 }
 
-// pybricks._common.LightMatrix.image
-STATIC mp_obj_t common_LightMatrix_image(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
+// pybricks._common.LightMatrix.icon
+STATIC mp_obj_t common_LightMatrix_icon(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
     PB_PARSE_ARGS_METHOD(n_args, pos_args, kw_args,
         common_LightMatrix_obj_t, self,
-        PB_ARG_REQUIRED(image));
+        PB_ARG_REQUIRED(icon));
 
     // Allocate and extract image data
     size_t size = pbio_light_matrix_get_size(self->light_matrix);
     common_LightMatrix__renew(self, 1);
-    common_LightMatrix_image__extract(image_in, size, self->data);
+    common_LightMatrix_icon__extract(icon_in, size, self->data);
 
-    // Display the image
+    // Display the icon
     pb_assert(pbio_light_matrix_set_image(self->light_matrix, self->data));
 
     return mp_const_none;
 }
-STATIC MP_DEFINE_CONST_FUN_OBJ_KW(common_LightMatrix_image_obj, 1, common_LightMatrix_image);
+STATIC MP_DEFINE_CONST_FUN_OBJ_KW(common_LightMatrix_icon_obj, 1, common_LightMatrix_icon);
 
 // pybricks._common.LightMatrix.on
 STATIC mp_obj_t common_LightMatrix_on(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
@@ -220,16 +220,16 @@ STATIC MP_DEFINE_CONST_FUN_OBJ_KW(common_LightMatrix_number_obj, 1, common_Light
 STATIC mp_obj_t common_LightMatrix_animate(size_t n_args, const mp_obj_t *pos_args, mp_map_t *kw_args) {
     PB_PARSE_ARGS_METHOD(n_args, pos_args, kw_args,
         common_LightMatrix_obj_t, self,
-        PB_ARG_REQUIRED(images),
+        PB_ARG_REQUIRED(icons),
         PB_ARG_REQUIRED(interval));
 
     // Time between frames
     mp_int_t interval = pb_obj_get_int(interval_in);
 
-    // Unpack the list of images
-    mp_obj_t *image_objs;
+    // Unpack the list of icons
+    mp_obj_t *icon_objs;
     size_t n;
-    mp_obj_get_array(images_in, &n, &image_objs);
+    mp_obj_get_array(icons_in, &n, &icon_objs);
     if (n > UINT8_MAX || n < 2) {
         pb_assert(PBIO_ERROR_INVALID_ARG);
     }
@@ -240,7 +240,7 @@ STATIC mp_obj_t common_LightMatrix_animate(size_t n_args, const mp_obj_t *pos_ar
 
     // Extract animation data
     for (uint8_t i = 0; i < n; i++) {
-        common_LightMatrix_image__extract(image_objs[i], size, self->data + size * size * i);
+        common_LightMatrix_icon__extract(icon_objs[i], size, self->data + size * size * i);
     }
 
     // Activate the animation
@@ -304,7 +304,7 @@ STATIC MP_DEFINE_CONST_FUN_OBJ_KW(common_LightMatrix_text_obj, 1, common_LightMa
 // dir(pybricks.builtins.LightMatrix)
 STATIC const mp_rom_map_elem_t common_LightMatrix_locals_dict_table[] = {
     { MP_ROM_QSTR(MP_QSTR_char),            MP_ROM_PTR(&common_LightMatrix_char_obj)            },
-    { MP_ROM_QSTR(MP_QSTR_image),           MP_ROM_PTR(&common_LightMatrix_image_obj)           },
+    { MP_ROM_QSTR(MP_QSTR_icon),            MP_ROM_PTR(&common_LightMatrix_icon_obj)            },
     { MP_ROM_QSTR(MP_QSTR_number),          MP_ROM_PTR(&common_LightMatrix_number_obj)          },
     { MP_ROM_QSTR(MP_QSTR_off),             MP_ROM_PTR(&common_LightMatrix_off_obj)             },
     { MP_ROM_QSTR(MP_QSTR_on),              MP_ROM_PTR(&common_LightMatrix_on_obj)              },